Full client consolidation (#184)
* created test client trait to remove duplicate code

* AMQPClient has been removed. EventHubsClientWrapper is the only client implementation

* cleanup and fixing minor issue causing failing tests
