+++ core/src/main/scala/org/apache/spark/sql/streaming/eventhubs/EventHubsSourceProvider.scala
@@ -17,7 +17,7 @@
 
 package org.apache.spark.sql.streaming.eventhubs
 
-import org.apache.spark.eventhubscommon.client.{ AMQPEventHubsClient, EventHubsClientWrapper }
+import org.apache.spark.eventhubscommon.client.EventHubsClientWrapper
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.SQLContext
 import org.apache.spark.sql.execution.streaming.Source
@@ -48,7 +48,7 @@ private[sql] class EventHubsSourceProvider
     new EventHubsSource(sqlContext,
                         parameters,
                         EventHubsClientWrapper.apply,
-                        AMQPEventHubsClient.apply)
+                        EventHubsClientWrapper.apply)
   }
 }
