+++ core/src/main/scala/org/apache/spark/streaming/eventhubs/EventHubsUtils.scala
@@ -18,11 +18,7 @@ package org.apache.spark.streaming.eventhubs
 
 import com.microsoft.azure.eventhubs.EventData
 import org.apache.spark.SparkConf
-import org.apache.spark.eventhubscommon.client.{
-  AMQPEventHubsClient,
-  Client,
-  EventHubsClientWrapper
-}
+import org.apache.spark.eventhubscommon.client.{ Client, EventHubsClientWrapper }
 import org.apache.spark.streaming.StreamingContext
 
 object EventHubsUtils {
@@ -56,7 +52,7 @@ object EventHubsUtils {
                                               progressDir,
                                               eventParams,
                                               EventHubsClientWrapper.apply,
-                                              AMQPEventHubsClient.apply)
+                                              EventHubsClientWrapper.apply)
     newStream
   }
