+++ core/src/test/scala/org/apache/spark/streaming/eventhubs/EventHubDirectDStreamSuite.scala
@@ -20,11 +20,7 @@ package org.apache.spark.streaming.eventhubs
 import org.mockito.{ Matchers, Mockito }
 import org.scalatest.mock.MockitoSugar
 import org.apache.spark.eventhubscommon.{ EventHubNameAndPartition, OffsetRecord }
-import org.apache.spark.eventhubscommon.client.{
-  AMQPEventHubsClient,
-  Client,
-  EventHubsClientWrapper
-}
+import org.apache.spark.eventhubscommon.client.{ Client, EventHubsClientWrapper }
 import org.apache.spark.rdd.RDD
 import org.apache.spark.streaming.{ Duration, Seconds, Time }
 
@@ -51,7 +47,7 @@ class EventHubDirectDStreamSuite extends EventHubTestSuiteBase with MockitoSugar
                                               progressRootPath.toString,
                                               Map("eh1" -> eventhubParameters),
                                               EventHubsClientWrapper.apply,
-                                              AMQPEventHubsClient.apply)
+                                              EventHubsClientWrapper.apply)
     val eventHubClientMock = mock[Client]
     val ehNameToClient = mutable.HashMap("eh1" -> eventHubClientMock)
 
@@ -71,7 +67,7 @@ class EventHubDirectDStreamSuite extends EventHubTestSuiteBase with MockitoSugar
                                               progressRootPath.toString,
                                               Map("eh1" -> eventhubParameters),
                                               EventHubsClientWrapper.apply,
-                                              AMQPEventHubsClient.apply)
+                                              EventHubsClientWrapper.apply)
     val eventHubClientMock = mock[Client]
     val ehNameToClient = mutable.HashMap("eh1" -> eventHubClientMock)
