+++ core/src/test/scala/org/apache/spark/streaming/eventhubs/ProgressTrackingAndCheckpointSuite.scala
@@ -621,72 +621,6 @@ class ProgressTrackingAndCheckpointSuite
     )
   }
 
-  test("progress files are clean up correctly with a fragile rest endpoint") {
-    val input = Seq(Seq(1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
-                    Seq(4, 5, 6, 7, 8, 9, 10, 1, 2, 3),
-                    Seq(7, 8, 9, 1, 2, 3, 4, 5, 6, 7))
-    val expectedOutputBeforeRestart =
-      Seq(Seq(2, 3, 5, 6, 8, 9), Seq(4, 5, 7, 8, 10, 2), Seq(6, 7, 9, 10, 3, 4))
-    // the order of the output should looks like there is no issue, because we reuse the fetched
-    // highest offset
-    val expectedOutputAfterRestart = Seq(Seq(6, 7, 9, 10, 3, 4),
-                                         Seq(8, 9, 11, 2, 5, 6),
-                                         Seq(10, 11, 3, 4, 7, 8),
-                                         Seq(),
-                                         Seq(),
-                                         Seq())
-
-    // ugly stuff to make things serializable
-    FragileEventHubClient.numBatchesBeforeCrashedEndpoint = 3
-    FragileEventHubClient.lastBatchWhenEndpointCrashed = 6
-    FragileEventHubClient.latestRecords = Map(EventHubNameAndPartition("eh1", 0) -> (9L, 9L),
-                                              EventHubNameAndPartition("eh1", 1) -> (9L, 9L),
-                                              EventHubNameAndPartition("eh1", 2) -> (9L, 9L))
-
-    testFragileStream(
-      input,
-      eventhubsParams = Map[String, Map[String, String]](
-        "eh1" -> Map(
-          "eventhubs.partition.count" -> "3",
-          "eventhubs.maxRate" -> "2",
-          "eventhubs.name" -> "eh1",
-          "eventhubs.namespace" -> "namespace",
-          "eventhubs.policyname" -> "policyname",
-          "eventhubs.policykey" -> "policykey"
-        )
-      ),
-      expectedOffsetsAndSeqs = Map(
-        eventhubNamespace ->
-          OffsetRecord(2000L,
-                       Map(EventHubNameAndPartition("eh1", 0) -> (3L, 3L),
-                           EventHubNameAndPartition("eh1", 1) -> (3L, 3L),
-                           EventHubNameAndPartition("eh1", 2) -> (3L, 3L)))),
-      operation = (inputDStream: EventHubDirectDStream) =>
-        inputDStream.map(eventData => eventData.getProperties.get("output").asInstanceOf[Int] + 1),
-      expectedOutput = expectedOutputBeforeRestart
-    )
-
-    val currentCheckpointDirectory = ssc.checkpointDir
-    ssc.stop()
-    reset()
-
-    ssc = new StreamingContext(currentCheckpointDirectory)
-
-    runStreamsWithEventHubInput(ssc,
-                                expectedOutputAfterRestart.length - 1,
-                                expectedOutputAfterRestart,
-                                useSet = true)
-
-    testProgressTracker(
-      eventhubNamespace,
-      expectedOffsetsAndSeqs = OffsetRecord(8000L,
-                                            Map(EventHubNameAndPartition("eh1", 0) -> (9L, 9L),
-                                                EventHubNameAndPartition("eh1", 1) -> (9L, 9L),
-                                                EventHubNameAndPartition("eh1", 2) -> (9L, 9L))),
-      9000L
-    )
-  }
-
   test("offset type is saved and recovered correctly from checkpoint") {
     val input = Seq(Seq(1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
                     Seq(4, 5, 6, 7, 8, 9, 10, 1, 2, 3),
